requirements.txt
langchain
langchain-google-genai
python-dotenv
pydantic

# Add specific versions if needed, e.g.:
# langchain==0.1.16
# langchain-google-genai==0.1.3
# python-dotenv==1.0.1
# pydantic==2.7.1
